Start groupby().mean() implementation.
