+++ core/dataset_operations_common.h
@@ -1,0 +1,67 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
+/// @file
+/// @author Simon Heybrock
+#ifndef SCIPP_CORE_DATASET_OPERATIONS_COMMON_H
+#define SCIPP_CORE_DATASET_OPERATIONS_COMMON_H
+
+namespace scipp::core {
+
+template <bool ApplyToData, class Func, class... Args>
+DataArray apply_and_drop_dim_impl(const DataConstProxy &a, Func func,
+                                  const Dim dim, Args &&... args) {
+  std::map<Dim, Variable> coords;
+  for (auto &&[d, coord] : a.coords())
+    if (d != dim)
+      coords.emplace(d, coord);
+
+  std::map<std::string, Variable> labels;
+  for (auto &&[name, label] : a.labels())
+    if (label.dims().inner() != dim)
+      labels.emplace(name, label);
+
+  std::map<std::string, Variable> attrs;
+  for (auto &&[name, attr] : a.attrs())
+    if (attr.dims().inner() != dim)
+      attrs.emplace(name, attr);
+
+  std::map<std::string, Variable> masks;
+  for (auto &&[name, mask] : a.masks())
+    if (mask.dims().inner() != dim)
+      masks.emplace(name, mask);
+
+  if constexpr (ApplyToData)
+    return DataArray(func(a.data(), dim, std::forward<Args>(args)...),
+                     std::move(coords), std::move(labels), std::move(masks),
+                     std::move(attrs), a.name());
+  else
+    return DataArray(func(a, dim, std::forward<Args>(args)...),
+                     std::move(coords), std::move(labels), std::move(masks),
+                     std::move(attrs), a.name());
+}
+
+template <class Func, class... Args>
+DataArray apply_to_data_and_drop_dim(const DataConstProxy &a, Func func,
+                                     const Dim dim, Args &&... args) {
+  return apply_and_drop_dim_impl<true>(a, func, dim,
+                                       std::forward<Args>(args)...);
+}
+
+template <class Func, class... Args>
+DataArray apply_and_drop_dim(const DataConstProxy &a, Func func, const Dim dim,
+                             Args &&... args) {
+  return apply_and_drop_dim_impl<false>(a, func, dim,
+                                        std::forward<Args>(args)...);
+}
+
+template <class Func, class... Args>
+Dataset apply_to_items(const DatasetConstProxy &d, Func func, Args &&... args) {
+  Dataset result;
+  for (const auto &[name, data] : d)
+    result.setData(name, func(data, std::forward<Args>(args)...));
+  return result;
+}
+
+} // namespace scipp::core
+
+#endif // SCIPP_CORE_DATASET_OPERATIONS_COMMON_H
